Refactor make_piece_cid to be in test utils
