+++ actors/runtime/Cargo.toml
@@ -31,6 +31,7 @@ fvm_sdk = { version = "0.6.0", optional = true }
 blake2b_simd = "1.0"
 fvm_ipld_blockstore = { version = "0.1" }
 fvm_ipld_encoding = "0.1.0"
+multihash = { version = "0.16.1", default-features = false }
 
 [dev-dependencies]
 derive_builder = "0.10.2"
